inpainting_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
diffusers==0.30.0
transformers==4.44.0
accelerate==0.33.0
pillow==10.4.0